Moved funcs for metric to sysinfo.py
