+++ agents/scripts/environment/dummy.env
@@ -1,0 +1,15 @@
+DEBIAN_FRONTEND=noninteractive
+
+INSTALL_FFI=false
+INSTALL_NCURSES=false
+INSTALL_GDBM=false
+INSTALL_OPEN_SSL=false
+INSTALL_READLINE=false
+INSTALL_SQLITE=false
+
+PYTHON_DIR=/opt/python
+
+BUILD_ZLIB_FROM_SRC=true
+
+INSTALL_CORO=true
+INSTALL_PYTEST_XDIST=false
